In this step, we're copy-pasting the business logic to a function in a new file called `useMessageBoardModel`.
This function is a so-called "custom hook". A "hook" is any function in a React app that starts with "use", e.g. `useState`. In turn, a custom hook, is any function in a React app that starts with "use" and uses other hooks like `useState` internally.

The benefit of this refactoring is that we can now separate business logic from markup, also allowing us to potentially test this code in isolation.