Setup
To practice working with effects, we'll be building a virtual version of a [Theremin](https://www.youtube.com/watch?v=K6KbEnGnymk). It's considered the first electronic instrument that was invented and works by placing both hands within an electromagnetic field, controlling the pitch (aka. "frequency") with the left hand, while controlling the volume (aka. "gain") with the right hand.

In our virtual version, we'll use the mouse position to control the pitch as well as the volume, with the x-axis corresponding to the pitch and the y-axis corresponding to the volume. To activate the theremin, the mouse needs to be pressed—releasing the mouse will stop the theremin.

This is an interesting example of working with effects because we have to synchronize mouse events into our app state, followed by synchronizing app state into sound. To output sound, we'll use the [Web Audio API](https://developer.mozilla.org/en-US/docs/Web/API/Web_Audio_API), which can only be controlled imperatively. By wrapping it into `useEffect`, we can integrate it with the declarative nature of React. To make working with the Web Audio API slightly easier, we'll use a provided `Sound` class that acts as a convenience wrapper.