Synchronizing mouse events into state
To get started, we'll implement a custom hook that returns a boolean, indicating whether the mouse is currently "down". We're using this boolean to render emojis, indicating if sound is currently playing (i.e. the party has started) or not.

To implement the custom hook, we'll register a global event listener for `mousedown` and `mouseup` and update an internal state accordingly.

Note that we can move this hook to a separate file, but we'll leave it here for now for the sake of simplicity of this example.