Playing a sound
As the next step, we'll add another custom hook that will be responsible for syncing the `isActive` state to an imperative instance of `Sound`. By looking at how this hook is called, we immediately notice that it's "syncing out" since it doesn't return any value. The only argument that it accepts is called `isActive`, indicating wether a sound is played.

By now, we're simply playing a sound once the mouse is clicked, but we don't consider the mouse position yet.