Adjusting pitch and volume based on the mouse position
We can use the existing `useTheremin` effect to add a global mouse move listener here that will adjust the pitch and volume of the sound. Again, we need to be sure to remove the event listener in the cleanup function.

By now, the theremin is already playable—give it a try!