Disco background
From here, we can add another effect that synchronizes our `isActive` state to the global background color of the document. It should do this based on an interval so that the background color changes every second to another new random color. Once `isActive` goes back to `false`, we'll reset the background color to the initial color.