Playing theremin with two hands
But a theremin is supposed to be played with two hands, right? Let's take this app a step further by requiring that the user presses the space bar while simultaneously using the mouse. `isActive` should now only be true when `isMouseDown` and `isSpacePressed` are both true. In case only one of them is set, we'll display a third emoji state, indicating that something is missing.

The implementation is similar to our mouse down handler: We register a global keydown and keyup handler and check whether the code of the keyboard event matches the code that was provided to the hook. If it does, we'll sync this event to a state within the hook.