Moving the mouse position into React state
If you've followed along closely, you might have noticed that there's one hook that currently shares two concerns: `useTheremin` both takes care of playing a sound and also adjusts the pitch and volume based on the mouse position.

We can synchronize the mouse position into React state via a `useMouseWindowPosition` hook, similarly to the other hooks we've created so far. However, adjusting the pitch based on a state change from this position is not straightforward. The reason is that we're creating a new `Sound` instance every time the effect in `useTheremin` is called. If we'd do this every time the mouse position changes, then the sound would be continuously interrupted since sound instances are immediately disposed and recreated after every mouse move.

There's an advanced pattern in React that can help with this situation though. If you remember from the React introduction, the `useRef` hook can be used to save references to DOM elements. However, the `useRef` hook can also be used to save arbitrary references to values that can be shared among event handlers and effects (see [Referencing a value with a ref](https://react.dev/reference/react/useRef#referencing-a-value-with-a-ref)).

We can use this to split the `useTheremin` hook into two parts: An effect that creates an instance of `Sound` once `isActive` is true, and another one that adjusts the pitch and volume based on the mouse position. The `Sound` instance can be saved to `soundRef` and retrieved via `soundRef.current`.

Again, this is an advanced pattern and something that is rarely needed. But if you find yourself in a situation where you need to share a value between event handlers or effects that doesn't necessarily need to be put into state (because it doesn't affect the render output of a component), then `useRef` can be a viable option.